Catch case where no ids are returned and ERP is overwrite; erroneous error was previously returned

+++ app/models/ca_data_importers.php
@@ -2167,10 +2167,10 @@ class ca_data_importers extends BundlableLabelableBaseModelWithAttributes {
 			if(!is_array($ids) || !sizeof($ids)) { $ids = [null]; }
 			foreach($ids as $id) {	
 				$t_subject->load($id);		
-				if(in_array($vs_existing_record_policy, ['overwrite_on_preferred_labels', 'overwrite_on_idno_and_preferred_labels', 'overwrite_on_idno'])) {
+				if($t_subject->isLoaded() && in_array($vs_existing_record_policy, ['overwrite_on_preferred_labels', 'overwrite_on_idno_and_preferred_labels', 'overwrite_on_idno'])) {
 					$t_subject->delete(true, ['hard' => true]);
 					if ($t_subject->numErrors()) {
-						$this->logImportError(_t('[%1] Could not delete existing record matched on label by policy %2', $vs_idno, $vs_existing_record_policy), $va_log_import_error_opts);
+						$this->logImportError(_t('[%1] Could not delete existing record matched on label by policy %2: %3', $vs_idno, $vs_existing_record_policy, join('; ', $t_subject->getErrors())), $va_log_import_error_opts);
 						// Don't stop?
 					} else {
 						if ($log_erp) { $o_log->logInfo(_t('[%1] Overwrote existing record matched on label by policy %2', $vs_idno, $vs_existing_record_policy)); }
